cloud/requirements_cloud.txt
polars==1.0.0
pandas==2.2.0
pyarrow==15.0.0
numpy==1.26.0
scikit-learn==1.4.0
pydantic-settings==2.1.0
pyyaml==6.0.1
tqdm==4.66.1
